+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+v3.2
+
+* react bridge: Removed `mapStatetoProps` function and use much simpler `props` which is just an Observable emitting
+  the props of the connected component
+* react bridge: The `store` argument passed as the `ConnectCallback` in the `connect()` function now calls .clone()
+  on the store internally and automatically calls .destroy() on the clone when the component is unmount. That way we
+  don't need custom cleanup logic inside `connect()`.
+
 v3.0
 
 * Removed `Action` type (use Subject and specify a name as 3rd argument to .addReducer() instead)